refactor: split odata_read_functions.cpp along its five seams
The file carried five responsibilities in ~2300 lines. Split it into one
translation unit per responsibility, moving whole functions verbatim:

- src/odata_read_functions.cpp  - table function registration, ODataReadBind
                                  and the bind-time named-parameter helpers
- src/odata_read_bind_data.cpp  - ODataReadBindData construction, factories,
                                  schema resolution, configuration accessors
- src/odata_read_scan_state.cpp - row buffer, progress tracker, paging and
                                  row emission, CloneForScan, scan callbacks
- src/odata_read_pushdown.cpp   - projection/filter pushdown wiring
- src/odata_type_resolver.cpp   - EDM -> DuckDB type resolution
- src/odata_error_mapping.cpp   - ODataErrorHandling

No declaration changed and no function body was rewritten, so the public
surface the ODP, Datasphere, SAC, Business Central, Dataverse and catalog
modules bind against is untouched. Row buffering stays with the scan state:
ODataRowBuffer and ODataProgressTracker have no caller outside that file,
so splitting them out would only produce two 35-line units.
