+++ src/odata_read_functions.cpp
@@ -13,1886 +13,17 @@
 #include "telemetry.hpp"
 #include "erpl_web_banner.hpp"
 
+// The odata_read() table function itself: its DuckDB registration, the bind
+// entry point and the bind-time named-parameter / expand-clause helpers.
+// The bind data it produces, the scan state that consumes it, pushdown wiring,
+// type resolution and error mapping live in the sibling odata_read_*.cpp,
+// odata_type_resolver.cpp and odata_error_mapping.cpp translation units.
+//
+// Declarations live in src/include/odata_read_functions.hpp.
+
 namespace erpl_web {
 
 using duckdb::PostHogTelemetry;
-
-namespace {
-
-using JsonDoc = duckdb_yyjson::yyjson_doc;
-using JsonValue = duckdb_yyjson::yyjson_val;
-
-struct JsonDocHandle {
-  explicit JsonDocHandle(JsonDoc *doc) : doc(doc) {}
-  ~JsonDocHandle() {
-    if (doc) {
-      duckdb_yyjson::yyjson_doc_free(doc);
-    }
-  }
-
-  JsonDocHandle(const JsonDocHandle &) = delete;
-  JsonDocHandle &operator=(const JsonDocHandle &) = delete;
-
-  JsonValue *Root() const {
-    return doc ? duckdb_yyjson::yyjson_doc_get_root(doc) : nullptr;
-  }
-
-  explicit operator bool() const { return doc != nullptr; }
-
-private:
-  JsonDoc *doc;
-};
-
-static bool IsODataMetadataField(const std::string &field_name) {
-  // OData V2 spells control information with a leading double underscore.
-  if (field_name == "__metadata" || field_name == "__deferred") {
-    return true;
-  }
-
-  // OData V4 (JSON Format, "Control Information") puts control information and
-  // annotations in names containing '@' - "@odata.etag", "@odata.id",
-  // "@odata.type", "@odata.editLink", and property-scoped forms such as
-  // "Price@odata.type". A structural property name is a SimpleIdentifier and can
-  // never contain '@', so this is an exact test rather than a heuristic.
-  //
-  // Treating "@odata.etag" as a column meant it was pushed into $select, which
-  // OData rejects outright: "Syntax error: character '@' is not valid at
-  // position 0 in '@odata'". Northwind V4 returns an etag on every entity, so
-  // the whole service was unreadable. See GitHub #108.
-  return field_name.find('@') != std::string::npos;
-}
-
-static bool IsODataV2DeferredNavigationProperty(duckdb_yyjson::yyjson_val *val) {
-  // In OData V2 JSON, navigation properties appear as {"__deferred": {"uri": "..."}}
-  // Skip these — they are not queryable columns.
-  if (!val || !duckdb_yyjson::yyjson_is_obj(val)) {
-    return false;
-  }
-  return duckdb_yyjson::yyjson_obj_get(val, "__deferred") != nullptr;
-}
-
-static bool AppendColumnNamesFromObject(
-    JsonValue *row_obj, std::vector<std::string> &column_names) {
-  if (!row_obj || !duckdb_yyjson::yyjson_is_obj(row_obj)) {
-    return false;
-  }
-
-  const auto initial_size = column_names.size();
-  duckdb_yyjson::yyjson_obj_iter iter;
-  duckdb_yyjson::yyjson_obj_iter_init(row_obj, &iter);
-
-  duckdb_yyjson::yyjson_val *key;
-  while ((key = duckdb_yyjson::yyjson_obj_iter_next(&iter))) {
-    if (!duckdb_yyjson::yyjson_is_str(key)) {
-      continue;
-    }
-    std::string column_name = duckdb_yyjson::yyjson_get_str(key);
-    if (IsODataMetadataField(column_name)) {
-      continue;
-    }
-    auto *val = duckdb_yyjson::yyjson_obj_iter_get_val(key);
-    if (IsODataV2DeferredNavigationProperty(val)) {
-      continue;  // Skip V2 navigation properties (__deferred links)
-    }
-    column_names.push_back(std::move(column_name));
-  }
-
-  return column_names.size() > initial_size;
-}
-
-static JsonValue *FirstObjectInArray(JsonValue *array_value) {
-  if (!array_value || !duckdb_yyjson::yyjson_is_arr(array_value)) {
-    return nullptr;
-  }
-
-  auto first_row = duckdb_yyjson::yyjson_arr_get_first(array_value);
-  return first_row && duckdb_yyjson::yyjson_is_obj(first_row) ? first_row
-                                                              : nullptr;
-}
-
-static JsonValue *FirstODataV4Row(JsonValue *root) {
-  return FirstObjectInArray(duckdb_yyjson::yyjson_obj_get(root, "value"));
-}
-
-static JsonValue *FirstODataV2Row(JsonValue *root) {
-  auto data_obj = duckdb_yyjson::yyjson_obj_get(root, "d");
-  if (!data_obj || !duckdb_yyjson::yyjson_is_obj(data_obj)) {
-    return nullptr;
-  }
-  return FirstObjectInArray(duckdb_yyjson::yyjson_obj_get(data_obj, "results"));
-}
-
-static bool AppendFirstODataRowColumnNames(
-    JsonValue *root, std::vector<std::string> &column_names) {
-  if (AppendColumnNamesFromObject(FirstODataV4Row(root), column_names)) {
-    return true;
-  }
-  if (AppendColumnNamesFromObject(FirstODataV2Row(root), column_names)) {
-    return true;
-  }
-
-  // Some callers already pass the first row object rather than the response
-  // envelope. Accept that shape to keep extraction local and metadata-free.
-  return AppendColumnNamesFromObject(root, column_names);
-}
-
-} // namespace
-
-void ODataReadBindData::SetNestedExpandPaths(
-    const std::vector<std::string> &nested_paths) {
-  if (data_extractor) {
-    data_extractor->SetNestedExpandPaths(nested_paths);
-  }
-}
-
-// ============================================================================
-// Helper Functions
-// ============================================================================
-
-// Helper function to extract column names from a JSON object
-std::vector<std::string>
-ExtractColumnNamesFromJson(const std::string &json_content) {
-  std::vector<std::string> column_names;
-
-  JsonDocHandle doc(duckdb_yyjson::yyjson_read(json_content.c_str(),
-                                               json_content.size(), 0));
-  if (!doc) {
-    return column_names;
-  }
-
-  auto root = doc.Root();
-  if (!root) {
-    return column_names;
-  }
-
-  if (!AppendFirstODataRowColumnNames(root, column_names)) {
-    ERPL_TRACE_WARN("ODATA_READ_BIND",
-                    "Could not find a row object for column-name extraction");
-  }
-
-  return column_names;
-}
-
-// ============================================================================
-// ODataProgressTracker Implementation
-// ============================================================================
-
-void ODataProgressTracker::SetTotalCount(uint64_t total) {
-    total_count_ = total;
-    has_total_ = true;
-}
-
-void ODataProgressTracker::IncrementRowsFetched(uint64_t count) {
-    rows_fetched_ += count;
-}
-
-double ODataProgressTracker::GetProgressFraction() const {
-    if (!has_total_ || total_count_ == 0) {
-        return -1.0; // unknown -> DuckDB will not show progress
-    }
-    double progress = 100.0 * (double)rows_fetched_ / (double)total_count_;
-  if (progress < 0.0) {
-    progress = 0.0;
-  }
-    return std::min(100.0, progress);
-}
-
-bool ODataProgressTracker::HasTotalCount() const { return has_total_; }
-
-uint64_t ODataProgressTracker::GetTotalCount() const { return total_count_; }
-
-uint64_t ODataProgressTracker::GetRowsFetched() const { return rows_fetched_; }
-
-void ODataProgressTracker::Reset() {
-    rows_fetched_ = 0;
-    total_count_ = 0;
-    has_total_ = false;
-}
-
-// ============================================================================
-// ODataRowBuffer Implementation
-// ============================================================================
-
-void ODataRowBuffer::AddRows(std::vector<std::vector<duckdb::Value>> rows) {
-    for (auto &row : rows) {
-        row_buffer_.emplace_back(std::move(row));
-    }
-}
-
-std::vector<duckdb::Value> ODataRowBuffer::GetNextRow() {
-    if (row_buffer_.empty()) {
-        return {};
-    }
-    auto row = std::move(row_buffer_.front());
-    row_buffer_.pop_front();
-    return row;
-}
-
-bool ODataRowBuffer::HasMoreRows() const { return !row_buffer_.empty(); }
-
-size_t ODataRowBuffer::Size() const { return row_buffer_.size(); }
-
-std::vector<std::vector<duckdb::Value>> ODataRowBuffer::CopyRows() const {
-    return std::vector<std::vector<duckdb::Value>>(row_buffer_.begin(),
-                                                   row_buffer_.end());
-}
-
-void ODataRowBuffer::Clear() { row_buffer_.clear(); }
-
-void ODataRowBuffer::SetHasNextPage(bool has_next) {
-    has_next_page_ = has_next;
-}
-
-bool ODataRowBuffer::HasNextPage() const { return has_next_page_; }
-
-// ============================================================================
-// ODataTypeResolver Implementation
-// ============================================================================
-
-ODataTypeResolver::ODataTypeResolver(
-    std::shared_ptr<ODataEntitySetClient> odata_client)
-    : odata_client(odata_client) {}
-
-duckdb::LogicalType ODataTypeResolver::ResolveNavigationPropertyType(
-    const std::string &property_name) const {
-    try {
-        auto entity_type = odata_client->GetCurrentEntityType();
-        
-    for (const auto &nav_prop : entity_type.navigation_properties) {
-            if (nav_prop.name == property_name) {
-                auto [is_collection, type_name] = ExtractCollectionType(nav_prop.type);
-                
-                duckdb::LogicalType base_type;
-                if (type_name.find("Edm.") == 0) {
-                    base_type = ConvertPrimitiveTypeString(type_name);
-                } else {
-                    base_type = ResolveEntityType(type_name);
-                }
-                
-                if (is_collection) {
-                    return duckdb::LogicalType::LIST(base_type);
-                }
-                
-                return base_type;
-            }
-        }
-  } catch (const std::exception &e) {
-    ERPL_TRACE_WARN("TYPE_RESOLVER",
-                    "Failed to resolve type for navigation property '" +
-                        property_name + "': " + e.what());
-    }
-    
-    return duckdb::LogicalTypeId::VARCHAR; // Default fallback
-}
-
-std::pair<bool, std::string> ODataTypeResolver::GetNavTargetFromCurrentEntity(
-    const std::string &nav_prop) const {
-  try {
-    auto entity_type = odata_client->GetCurrentEntityType();
-    for (const auto &np : entity_type.navigation_properties) {
-      if (np.name == nav_prop) {
-        auto [is_collection, type_name] = ExtractCollectionType(np.type);
-        return {is_collection, type_name};
-      }
-    }
-  } catch (...) {
-  }
-  return {false, std::string()};
-}
-
-std::pair<bool, std::string>
-ODataTypeResolver::GetNavTargetOnEntity(const std::string &entity_type_name,
-                                        const std::string &nav_prop) const {
-  try {
-    auto edmx = odata_client->GetMetadata();
-    auto type_variant = edmx.FindType(entity_type_name);
-    EntityType entity;
-    if (std::holds_alternative<EntityType>(type_variant)) {
-      entity = std::get<EntityType>(type_variant);
-    } else {
-      return {false, std::string()};
-    }
-    for (const auto &np : entity.navigation_properties) {
-      if (np.name == nav_prop) {
-        auto [is_collection, type_name] = ExtractCollectionType(np.type);
-        return {is_collection, type_name};
-      }
-    }
-  } catch (...) {
-  }
-  return {false, std::string()};
-}
-
-duckdb::LogicalType ODataTypeResolver::ResolveNavigationOnEntity(
-    const std::string &entity_type_name, const std::string &nav_prop) const {
-  auto [is_collection, type_name] =
-      GetNavTargetOnEntity(entity_type_name, nav_prop);
-  if (type_name.empty())
-    return duckdb::LogicalTypeId::VARCHAR;
-  duckdb::LogicalType base_type;
-  if (type_name.find("Edm.") == 0) {
-    base_type = ConvertPrimitiveTypeString(type_name);
-  } else {
-    base_type = ResolveEntityType(type_name);
-  }
-  if (is_collection) {
-    return duckdb::LogicalType::LIST(base_type);
-  }
-  return base_type;
-}
-
-duckdb::LogicalType ODataTypeResolver::ConvertPrimitiveTypeString(
-    const std::string &type_name) const {
-    return DuckTypeConverter::ConvertEdmPrimitiveStringToLogicalType(type_name);
-}
-
-std::tuple<bool, std::string>
-ODataTypeResolver::ExtractCollectionType(const std::string &type_name) const {
-    std::regex collection_regex("Collection\\(([^\\)]+)\\)");
-    std::smatch match;
-
-    if (std::regex_search(type_name, match, collection_regex)) {
-        return std::make_tuple(true, match[1]);
-    } else {
-        return std::make_tuple(false, type_name);
-    }
-}
-
-duckdb::LogicalType
-ODataTypeResolver::ResolveEntityType(const std::string &type_name) const {
-    try {
-        auto edmx = odata_client->GetMetadata();
-        auto target_type = edmx.FindType(type_name);
-
-        // Use DuckTypeConverter to get proper DuckDB type
-        auto type_conv = DuckTypeConverter(edmx);
-        return std::visit(type_conv, target_type);
-    } catch (const std::exception &e) {
-        ERPL_TRACE_WARN("TYPE_RESOLVER", "Failed to resolve entity type '" +
-                                             type_name + "': " + e.what());
-        return duckdb::LogicalTypeId::VARCHAR;
-    }
-}
-
-duckdb::LogicalType
-ODataTypeResolver::ResolveComplexType(const std::string &type_name) const {
-    try {
-        auto edmx = odata_client->GetMetadata();
-        auto target_type = edmx.FindType(type_name);
-
-        // Use DuckTypeConverter to get proper DuckDB type
-        auto type_conv = DuckTypeConverter(edmx);
-        return std::visit(type_conv, target_type);
-    } catch (const std::exception &e) {
-        ERPL_TRACE_WARN("TYPE_RESOLVER", "Failed to resolve complex type '" +
-                                         type_name + "': " + e.what());
-        return duckdb::LogicalTypeId::VARCHAR;
-    }
-}
-
-// ============================================================================
-// ODataReadBindData Implementation
-// ============================================================================
-
-ODataReadBindData::ODataReadBindData(
-    std::shared_ptr<ODataEntitySetClient> odata_client)
-    : odata_client(odata_client) {
-    InitializeComponents();
-}
-
-ODataReadBindData::ODataReadBindData(
-    std::shared_ptr<ODataEntitySetClient> odata_client,
-    bool defer_initialization)
-    : odata_client(odata_client) {
-  // Don't call InitializeComponents() - let the factory method do it with the
-  // correct mode
-}
-
-void ODataReadBindData::InitializeComponents(bool service_root_mode) {
-    predicate_pushdown_helper = nullptr;
-
-  // The failure log is created first and owned here: page responses and the
-  // row buffer are replaced during pagination, the log must not be.
-  conversion_failure_log = std::make_shared<ConversionFailureLog>();
-
-  // Only create entity-set specific components if not in service root mode
-  if (!service_root_mode) {
-    data_extractor = std::make_shared<ODataDataExtractor>(odata_client);
-    data_extractor->SetConversionFailureLog(conversion_failure_log);
-    type_resolver = std::make_shared<ODataTypeResolver>(odata_client);
-  } else {
-    // In service root mode, these components are not needed and would trigger
-    // metadata fetching
-    data_extractor = nullptr;
-    type_resolver = nullptr;
-    ERPL_TRACE_INFO("ODATA_READ_BIND",
-                    "Skipping data_extractor and type_resolver creation in "
-                    "service root mode");
-  }
-
-    progress_tracker = std::make_shared<ODataProgressTracker>();
-    row_buffer = std::make_shared<ODataRowBuffer>();
-}
-
-ODataReadBindData::~ODataReadBindData() {
-  // Safety net: a query with a LIMIT may be torn down before the scan reaches
-  // its terminal call, and the user should still learn about bad values.
-  try {
-    ReportConversionFailures();
-  } catch (...) {
-    // Never let a destructor throw.
-  }
-}
-
-void ODataReadBindData::SetStrictTyping(bool strict) {
-  strict_typing_ = strict;
-  if (!conversion_failure_log) {
-    conversion_failure_log = std::make_shared<ConversionFailureLog>();
-  }
-  conversion_failure_log->SetStrictTyping(strict);
-  ERPL_TRACE_INFO("ODATA_READ_BIND",
-                  std::string("strict_typing set to ") +
-                      (strict ? "true" : "false"));
-}
-
-std::shared_ptr<ConversionFailureLog>
-ODataReadBindData::GetConversionFailureLog() const {
-  return conversion_failure_log;
-}
-
-void ODataReadBindData::ReportConversionFailures() {
-  if (conversion_failures_reported_ || !conversion_failure_log) {
-    return;
-  }
-  conversion_failures_reported_ = true;
-  conversion_failure_log->ReportAndDrain("odata_read");
-}
-
-// Helper methods for URL detection
-bool ODataReadBindData::IsDatasphereUrl(const std::string &entity_set_url) {
-    return entity_set_url.find("datasphere") != std::string::npos || 
-           entity_set_url.find("hcs.cloud.sap") != std::string::npos;
-}
-
-bool ODataReadBindData::IsODataV2Url(const std::string &entity_set_url) {
-    return entity_set_url.find("/V2/") != std::string::npos;
-}
-
-bool ODataReadBindData::ShouldUseDirectHttp(const std::string &entity_set_url) {
-    return IsDatasphereUrl(entity_set_url) || IsODataV2Url(entity_set_url);
-}
-
-duckdb::unique_ptr<ODataReadBindData> ODataReadBindData::FromEntitySetRoot(
-    const std::string &entity_set_url,
-    std::shared_ptr<HttpAuthParams> auth_params) {
-  // Create HTTP client with URL encoding disabled for OData (V2 & V4)
-  HttpParams http_params;
-  http_params.url_encode = false;
-  auto http_client = std::make_shared<HttpClient>(http_params);
-
-  // Ensure $format=json is present on the entity set URL; encode $filter via
-  // ODataUrlCodec
-  HttpUrl es_url(entity_set_url);
-  ODataUrlCodec::ensureJsonFormat(es_url);
-  auto odata_client =
-      std::make_shared<ODataEntitySetClient>(http_client, es_url, auth_params);
-    
-    // Determine OData version and approach
-    bool is_odata_v2_url = IsODataV2Url(entity_set_url);
-    bool use_direct_http = ShouldUseDirectHttp(entity_set_url);
-    
-    // Set OData version directly for OData v2 to skip metadata fetching
-    if (is_odata_v2_url) {
-        odata_client->SetODataVersionDirectly(ODataVersion::V2);
-    ERPL_TRACE_DEBUG(
-        "ODATA_READ_BIND",
-        "Set OData version to V2 directly to skip metadata fetching");
-    }
-    
-    std::vector<std::string> extracted_column_names;
-    
-    if (use_direct_http) {
-        // Make direct HTTP request to get first page for column extraction
-        if (IsDatasphereUrl(entity_set_url)) {
-      ERPL_TRACE_DEBUG("ODATA_READ_BIND",
-                       "Detected Datasphere URL, making direct HTTP request to "
-                       "get first data page for @odata.context");
-        } else {
-      ERPL_TRACE_DEBUG("ODATA_READ_BIND",
-                       "Detected OData v2 URL, making direct HTTP request to "
-                       "extract column names");
-        }
-        
-        try {
-      HttpRequest req(HttpMethod::GET, HttpUrl(es_url));
-            req.AuthHeadersFromParams(*auth_params);
-            req.headers["Accept"] = "application/json";
-            auto resp = http_client->SendRequest(req);
-            
-            if (resp && resp->Code() == 200) {
-        ERPL_TRACE_DEBUG(
-            "ODATA_READ_BIND",
-            "Successfully fetched first data page via direct HTTP");
-                
-                // Parse the response to extract @odata.context and column names
-                auto content = resp->Content();
-        ERPL_TRACE_DEBUG("ODATA_READ_BIND", "Parsing JSON response of " +
-                                                std::to_string(content.size()) +
-                                                " bytes");
-        JsonDocHandle doc(
-            duckdb_yyjson::yyjson_read(content.c_str(), content.size(), 0));
-                
-                if (doc) {
-          ERPL_TRACE_DEBUG("ODATA_READ_BIND",
-                           "Successfully parsed JSON document");
-                    auto root = doc.Root();
-                    ERPL_TRACE_DEBUG("ODATA_READ_BIND", "Got JSON root object");
-                    
-                    // Check for @odata.context (OData v4)
-          auto context_val =
-              duckdb_yyjson::yyjson_obj_get(root, "@odata.context");
-                    if (context_val && duckdb_yyjson::yyjson_is_str(context_val)) {
-                        auto context_url = duckdb_yyjson::yyjson_get_str(context_val);
-            ERPL_TRACE_DEBUG("ODATA_READ_BIND", "Extracted @odata.context: " +
-                                                    std::string(context_url));
-                        
-            // Store the metadata context URL in the OData client for future
-            // metadata requests
-                        std::string clean_context_url = context_url;
-                        auto hash_pos = clean_context_url.find('#');
-                        if (hash_pos != std::string::npos) {
-                            clean_context_url = clean_context_url.substr(0, hash_pos);
-                        }
-                        
-                        if (IsDatasphereUrl(entity_set_url)) {
-                            odata_client->SetMetadataContextUrl(clean_context_url);
-              ERPL_TRACE_DEBUG("ODATA_READ_BIND",
-                               "Stored metadata context URL in OData client: " +
-                                   clean_context_url);
-                            odata_client->SetEntitySetNameFromContextFragment(context_url);
-                        }
-                        
-                        // Parse OData V4 response format
-                        ParseODataV4Response(root, odata_client, extracted_column_names);
-                    } else {
-                        // Parse OData V2 response format
-                        ParseODataV2Response(root, odata_client, extracted_column_names);
-                    }
-                }
-            } else {
-        ERPL_TRACE_WARN("ODATA_READ_BIND",
-                        "Direct HTTP request failed with status: " +
-                            std::to_string(resp ? resp->Code() : 0));
-      }
-    } catch (const std::exception &e) {
-      ERPL_TRACE_WARN("ODATA_READ_BIND",
-                      "Failed to fetch first data page via direct HTTP: " +
-                          std::string(e.what()));
-            // Continue without first page - will fall back to conventional metadata
-        }
-    } else {
-    ERPL_TRACE_DEBUG(
-        "ODATA_READ_BIND",
-        "Standard OData v4 URL detected, using conventional metadata approach");
-    }
-    
-    auto bind_data = duckdb::make_uniq<ODataReadBindData>(odata_client);
-    
-    // If we successfully extracted column names, store them for later use
-    if (!extracted_column_names.empty()) {
-        bind_data->SetExtractedColumnNames(extracted_column_names);
-    }
-    
-    return bind_data;
-}
-
-bool ODataReadBindData::LooksLikeServiceRootUrl(const std::string &url) {
-  // Strict V2 service root detection: ends with .svc or .svc/
-  auto qpos = url.find('?');
-  std::string path_only = (qpos == std::string::npos) ? url : url.substr(0, qpos);
-  if (path_only.rfind("/$metadata") != std::string::npos) {
-    return false;
-  }
-  auto pos = path_only.rfind(".svc");
-  if (pos == std::string::npos) {
-    return false;
-  }
-  // Ensure there are no further path segments after .svc except optional '/'
-  size_t after = pos + 4; // length of ".svc"
-  if (after == path_only.size()) {
-    return true;
-  }
-  if (after + 1 == path_only.size() && path_only[after] == '/') {
-    return true;
-  }
-  return false;
-}
-
-duckdb::unique_ptr<ODataReadBindData> ODataReadBindData::FromServiceRoot(
-    const std::string &service_root_url,
-    std::shared_ptr<HttpAuthParams> auth_params) {
-  // Build minimal schema: name, kind, url (absolute)
-  // Reuse ODataServiceClient to fetch and parse the service document, then
-  // project rows locally.
-  HttpParams http_params;
-  http_params.url_encode = false;
-  auto http_client = std::make_shared<HttpClient>(http_params);
-
-  HttpUrl svc_url(service_root_url);
-  // Ensure JSON format for V2 roots
-  ODataUrlCodec::ensureJsonFormat(svc_url);
-
-  auto svc_client =
-      std::make_shared<ODataServiceClient>(http_client, svc_url, auth_params);
-  // Force V2 headers/parsing for classic service roots when URL indicates V2
-  try {
-    std::string url_str = service_root_url;
-    if (url_str.find("/V2/") != std::string::npos) {
-      svc_client->SetODataVersionDirectly(ODataVersion::V2);
-      ERPL_TRACE_DEBUG("ODATA_READ_BIND",
-                       "FromServiceRoot: forced OData version V2 based on URL");
-    }
-  } catch (...) {
-  }
-  // Try to detect/fetch
-  auto response = svc_client->Get();
-
-  // Build an entity-set client wrapper so we can reuse existing bind flow but
-  // override names/types via extracted list We'll populate
-  // extracted_column_names with our unified schema
-  auto bind_data = duckdb::make_uniq<ODataReadBindData>(
-      std::make_shared<ODataEntitySetClient>(http_client, svc_url,
-                                             auth_params));
-
-  // Unified schema (name, kind, url)
-  std::vector<std::string> names = {"name", "kind", "url"};
-  bind_data->SetExtractedColumnNames(names);
-
-  // Convert service entries to synthetic rows and buffer them
-  auto entries = response->EntitySets();
-  for (auto &ref : entries) {
-    ref.MergeWithBaseUrlIfRelative(svc_url);
-  }
-
-  // Build a display base URL without query parameters and with trailing slash
-  std::string display_base = service_root_url;
-  auto qpos_display = display_base.find('?');
-  if (qpos_display != std::string::npos) {
-    display_base = display_base.substr(0, qpos_display);
-  }
-  if (!display_base.empty() && display_base.back() != '/') {
-    display_base.push_back('/');
-  }
-
-  // Bootstrap row buffer and push rows
-  auto rows = std::vector<std::vector<duckdb::Value>>();
-  rows.reserve(entries.size());
-  for (const auto &ref : entries) {
-    // For service root listing, expose absolute URL by extending the base path
-    rows.push_back({duckdb::Value(ref.name), duckdb::Value("EntitySet"),
-                    duckdb::Value(display_base + ref.name)});
-  }
-
-  // Initialize internal components for buffering
-  bind_data->InitializeComponents();
-  bind_data->EnableServiceRootMode();
-  bind_data->row_buffer->AddRows(std::move(rows));
-  bind_data->row_buffer->SetHasNextPage(false);
-  bind_data->first_page_cached_ = true;
-
-  return bind_data;
-}
-
-// Factory pattern methods
-duckdb::unique_ptr<ODataReadBindData> ODataReadBindData::FromProbeResult(
-    const ODataClientFactory::ProbeResult &result) {
-  ERPL_TRACE_INFO(
-      "ODATA_READ_BIND",
-      std::string("Creating bind data from probe result - is_service_root: ") +
-          (result.is_service_root ? "true" : "false"));
-
-  if (result.is_service_root) {
-    ERPL_TRACE_INFO("ODATA_READ_BIND", "Taking service root path");
-    auto client = ODataClientFactory::CreateServiceClient(result);
-    return FromServiceClient(client, result.initial_content);
-  } else {
-    ERPL_TRACE_INFO("ODATA_READ_BIND", "Taking entity set path");
-    auto client = ODataClientFactory::CreateEntitySetClient(result);
-    return FromEntitySetClient(client, result.initial_content);
-  }
-}
-
-duckdb::unique_ptr<ODataReadBindData> ODataReadBindData::FromEntitySetClient(
-    std::shared_ptr<ODataEntitySetClient> client,
-    const std::string &initial_content) {
-  ERPL_TRACE_DEBUG("ODATA_READ_BIND",
-                   "Creating bind data from entity set client");
-
-  auto bind_data = duckdb::make_uniq<ODataReadBindData>(client, true);
-
-  // Initialize components for entity set mode (allows metadata fetching)
-  bind_data->InitializeComponents(false);
-
-  // If we have initial content, we can pre-populate the first page
-  if (!initial_content.empty()) {
-    // Parse the initial content to extract column names and types
-    try {
-      JsonDocHandle doc(duckdb_yyjson::yyjson_read(
-          initial_content.c_str(), initial_content.length(), 0));
-      if (doc) {
-        auto root = doc.Root();
-        if (root) {
-          std::vector<std::string> extracted_column_names;
-
-          // Determine OData version and parse accordingly
-          if (client->GetODataVersion() == ODataVersion::V4) {
-            ParseODataV4Response(root, client, extracted_column_names);
-          } else {
-            ParseODataV2Response(root, client, extracted_column_names);
-          }
-
-          bind_data->SetExtractedColumnNames(extracted_column_names);
-        }
-      }
-    } catch (...) {
-      // If parsing fails, we'll fall back to normal metadata-based approach
-      ERPL_TRACE_DEBUG(
-          "ODATA_READ_BIND",
-          "Failed to parse initial content, will use metadata approach");
-    }
-
-    // Buffer the rows from initial_content so the next FetchNextResult does
-    // not issue a redundant bare GET. Without this, callers that hand a
-    // pre-fetched page in (e.g. the ODP streaming path) trigger
-    // PrefetchFirstPage → odata_client->Get() with no query string, which
-    // SAP ODP answers with the entire dataset, inflating results N×.
-    try {
-      auto synthetic = std::make_shared<ODataEntitySetResponse>(
-          std::make_unique<HttpResponse>(HttpMethod::GET, HttpUrl(client->Url()),
-                                         200, "application/json", initial_content),
-          client->GetODataVersion());
-      bind_data->BufferFirstPageFromResponse(synthetic);
-    } catch (const std::exception &e) {
-      ERPL_TRACE_DEBUG(
-          "ODATA_READ_BIND",
-          std::string("Failed to buffer initial content; "
-                      "PrefetchFirstPage will refetch: ") +
-              e.what());
-    }
-  }
-
-  return bind_data;
-}
-
-duckdb::unique_ptr<ODataReadBindData>
-ODataReadBindData::FromServiceClient(std::shared_ptr<ODataServiceClient> client,
-                                     const std::string &initial_content) {
-  ERPL_TRACE_INFO("ODATA_READ_BIND", "Creating bind data from service client");
-
-  // Create a minimal stub entity set client for compatibility with existing
-  // bind data structure This client will never be used for actual HTTP requests
-  // in service root mode
-  HttpParams http_params;
-  http_params.url_encode = false;
-  auto http_client = std::make_shared<HttpClient>(http_params);
-
-  // Use a dummy URL to avoid any metadata resolution attempts
-  auto stub_client = std::make_shared<ODataEntitySetClient>(
-      http_client, HttpUrl("http://localhost/stub"), nullptr);
-  stub_client->SetODataVersionDirectly(client->GetODataVersion());
-
-  auto bind_data = duckdb::make_uniq<ODataReadBindData>(stub_client, true);
-
-  // Always enable service root mode for service clients
-  bind_data->EnableServiceRootMode();
-
-  // Set unified schema (name, kind, url) BEFORE initializing components
-  // This ensures GetResultNames() never falls back to calling the client
-  std::vector<std::string> names = {"name", "kind", "url"};
-  bind_data->SetExtractedColumnNames(names);
-
-  // Initialize components with service root mode (this prevents metadata
-  // fetching)
-  bind_data->InitializeComponents(true);
-  ERPL_TRACE_DEBUG("ODATA_READ_BIND",
-                   "FromServiceClient: service root mode enabled");
-
-  // If we have initial content, parse it via ODataServiceResponse (managed
-  // yyjson lifecycle)
-  if (!initial_content.empty()) {
-    try {
-      auto temp_response = std::make_shared<ODataServiceResponse>(
-          std::make_unique<HttpResponse>(HttpMethod::GET, HttpUrl(""), 200,
-                                         "application/json", initial_content),
-          client->GetODataVersion());
-
-      auto entries = temp_response->EntitySets();
-
-      // Convert service entries to synthetic rows and buffer them with absolute URLs
-      auto rows = std::vector<std::vector<duckdb::Value>>();
-      rows.reserve(entries.size());
-
-      // Derive display base from the service client URL
-      std::string display_base = client->Url();
-      auto qpos_display = display_base.find('?');
-      if (qpos_display != std::string::npos) {
-        display_base = display_base.substr(0, qpos_display);
-      }
-      if (!display_base.empty() && display_base.back() != '/') {
-        display_base.push_back('/');
-      }
-
-      for (const auto &ref : entries) {
-        std::string kind = "EntitySet"; // could be extended later for Singletons, etc.
-        rows.push_back({duckdb::Value(ref.name), duckdb::Value(kind),
-                        duckdb::Value(display_base + ref.name)});
-      }
-
-      // Buffer rows (components already initialized in service root mode)
-      bind_data->row_buffer->AddRows(std::move(rows));
-      bind_data->row_buffer->SetHasNextPage(false);
-      bind_data->first_page_cached_ = true;
-    } catch (...) {
-      ERPL_TRACE_DEBUG("ODATA_READ_BIND",
-                       "Failed to parse service root initial content; "
-                       "continuing without prebuffer");
-    }
-  } else {
-    // Fallback: fetch now and buffer entries
-    try {
-      auto response = client->Get();
-      auto entries = response->EntitySets();
-
-      auto rows = std::vector<std::vector<duckdb::Value>>();
-      rows.reserve(entries.size());
-      for (const auto &ref : entries) {
-        std::string kind = "EntitySet";
-        rows.push_back({duckdb::Value(ref.name), duckdb::Value(kind),
-                        duckdb::Value(ref.name)});
-      }
-
-      bind_data->row_buffer->AddRows(std::move(rows));
-      bind_data->row_buffer->SetHasNextPage(false);
-      bind_data->first_page_cached_ = true;
-    } catch (...) {
-      ERPL_TRACE_DEBUG("ODATA_READ_BIND",
-                       "FromServiceClient fallback fetch failed");
-    }
-    }
-    
-    return bind_data;
-}
-
-// Helper methods for parsing OData responses
-void ODataReadBindData::ParseODataV4Response(
-    duckdb_yyjson::yyjson_val *root,
-    std::shared_ptr<ODataEntitySetClient> odata_client,
-    std::vector<std::string> &extracted_column_names) {
-  (void)odata_client;
-  ERPL_TRACE_DEBUG("ODATA_READ_BIND", "Parsing OData v4 response format");
-  if (AppendColumnNamesFromObject(FirstODataV4Row(root),
-                                  extracted_column_names)) {
-    ERPL_TRACE_DEBUG("ODATA_READ_BIND",
-                     "Extracted column names from OData v4 first row");
-    return;
-  }
-
-  ERPL_TRACE_WARN("ODATA_READ_BIND",
-                  "Could not extract column names from OData v4 response");
-}
-
-void ODataReadBindData::ParseODataV2Response(
-    duckdb_yyjson::yyjson_val *root,
-    std::shared_ptr<ODataEntitySetClient> odata_client,
-    std::vector<std::string> &extracted_column_names) {
-  (void)odata_client;
-  ERPL_TRACE_DEBUG("ODATA_READ_BIND",
-                   "No @odata.context found, trying OData v2 format");
-  if (AppendColumnNamesFromObject(FirstODataV2Row(root),
-                                  extracted_column_names)) {
-    ERPL_TRACE_DEBUG("ODATA_READ_BIND",
-                     "Extracted column names from OData v2 first row");
-    return;
-  }
-
-  ERPL_TRACE_WARN("ODATA_READ_BIND",
-                  "Could not extract column names from OData v2 response");
-}
-
-const std::vector<std::string> &ODataReadBindData::MetadataColumnNames() {
-  if (!all_result_names.empty() || service_root_mode_ || !odata_client) {
-    return all_result_names;
-  }
-  try {
-    ERPL_TRACE_INFO("ODATA_READ_BIND",
-                    "Calling odata_client->GetResultNames() for metadata");
-    all_result_names = odata_client->GetResultNames();
-  } catch (const std::exception &e) {
-    ERPL_TRACE_WARN("ODATA_READ_BIND",
-                    std::string("Metadata column names unavailable: ") + e.what());
-  }
-  return all_result_names;
-}
-
-const std::vector<duckdb::LogicalType> &ODataReadBindData::MetadataColumnTypes() {
-  if (!all_result_types.empty() || service_root_mode_ || !odata_client) {
-    return all_result_types;
-  }
-  try {
-    ERPL_TRACE_INFO("ODATA_READ_BIND",
-                    "Calling odata_client->GetResultTypes() for metadata");
-    all_result_types = odata_client->GetResultTypes();
-  } catch (const std::exception &e) {
-    ERPL_TRACE_WARN("ODATA_READ_BIND",
-                    std::string("Metadata column types unavailable: ") + e.what());
-  }
-  return all_result_types;
-}
-
-std::vector<std::string> ODataReadBindData::ReconcileSchemaOrder(
-    const std::vector<std::string> &metadata_names,
-    const std::vector<std::string> &json_names) {
-  // Metadata (EDMX) order wins: the DuckDB catalog entry for an ATTACHed
-  // service declares its columns from EDMX, and DuckDB indexes the scan's
-  // column_ids against that declaration. Deriving the scan schema from JSON
-  // key order instead put values under the wrong headers (GitHub #88).
-  if (metadata_names.empty()) {
-    return json_names;
-  }
-  if (json_names.empty()) {
-    return metadata_names;
-  }
-
-  std::vector<std::string> reconciled = metadata_names;
-  std::unordered_set<std::string> known(metadata_names.begin(),
-                                        metadata_names.end());
-  for (const auto &json_name : json_names) {
-    if (known.insert(json_name).second) {
-      // Present in the payload but not declared in EDMX. Keep it, but only
-      // after every declared column so the metadata prefix stays aligned with
-      // the catalog.
-      reconciled.push_back(json_name);
-    }
-  }
-  return reconciled;
-}
-
-void ODataReadBindData::EnsureBaseSchemaResolved() {
-  if (base_schema_resolved_) {
-    return;
-  }
-
-  if (service_root_mode_) {
-    base_result_names = extracted_column_names.empty()
-                            ? std::vector<std::string>{"name", "kind", "url"}
-                            : extracted_column_names;
-    base_result_types.assign(
-        base_result_names.size(),
-        duckdb::LogicalType(duckdb::LogicalTypeId::VARCHAR));
-    base_schema_resolved_ = true;
-    return;
-  }
-
-  const auto metadata_names = MetadataColumnNames();
-  const auto metadata_types = MetadataColumnTypes();
-
-  base_result_names = ReconcileSchemaOrder(metadata_names, extracted_column_names);
-
-  base_result_types.clear();
-  base_result_types.reserve(base_result_names.size());
-  for (const auto &column_name : base_result_names) {
-    auto it = std::find(metadata_names.begin(), metadata_names.end(), column_name);
-    if (it != metadata_names.end()) {
-      const auto metadata_index =
-          static_cast<size_t>(std::distance(metadata_names.begin(), it));
-      if (metadata_index < metadata_types.size()) {
-        base_result_types.push_back(metadata_types[metadata_index]);
-        continue;
-      }
-    }
-    ERPL_TRACE_WARN("ODATA_READ_BIND",
-                    "Column '" + column_name +
-                        "' has no declared metadata type, using VARCHAR");
-    base_result_types.push_back(
-        duckdb::LogicalType(duckdb::LogicalTypeId::VARCHAR));
-  }
-
-  ERPL_TRACE_INFO("ODATA_READ_BIND",
-                  duckdb::StringUtil::Format(
-                      "Resolved base schema: %d columns (%d from metadata)",
-                      (int)base_result_names.size(), (int)metadata_names.size()));
-
-  base_schema_resolved_ = true;
-}
-
-std::vector<std::string> ODataReadBindData::GetResultNames(bool all_columns) {
-  EnsureBaseSchemaResolved();
-
-  std::vector<std::string> base_names = base_result_names;
-
-  // Add expanded data columns if we have any, avoiding duplicates
-  if (HasExpandedData()) {
-    for (const auto &expand_name : data_extractor->GetExpandedDataSchema()) {
-      if (std::find(base_names.begin(), base_names.end(), expand_name) ==
-          base_names.end()) {
-        ERPL_TRACE_DEBUG("ODATA_READ_BIND",
-                         "Adding expanded column: " + expand_name);
-        base_names.push_back(expand_name);
-      }
-    }
-  }
-
-  if (all_columns || active_column_ids.empty()) {
-    return base_names;
-  }
-
-  std::vector<std::string> active_result_names;
-  for (auto &column_id : active_column_ids) {
-    if (duckdb::IsRowIdColumnId(column_id)) {
-      continue;
-    }
-    if (column_id < base_names.size()) {
-      active_result_names.push_back(base_names[column_id]);
-    }
-  }
-
-  return active_result_names;
-}
-
-std::vector<duckdb::LogicalType>
-ODataReadBindData::GetResultTypes(bool all_columns) {
-  EnsureBaseSchemaResolved();
-
-  std::vector<duckdb::LogicalType> combined_types = base_result_types;
-  const auto &base_names_local = base_result_names;
-
-  // Add/merge expanded data types if we have any, keeping them aligned with the
-  // names GetResultNames() produces.
-  if (HasExpandedData()) {
-    const auto &exp_schema = data_extractor->GetExpandedDataSchema();
-    const auto &exp_types = data_extractor->GetExpandedDataTypes();
-    for (size_t i = 0; i < exp_schema.size(); ++i) {
-      const auto &exp_name = exp_schema[i];
-      duckdb::LogicalType exp_type =
-          (i < exp_types.size())
-              ? exp_types[i]
-              : duckdb::LogicalType(duckdb::LogicalTypeId::VARCHAR);
-
-      auto it =
-          std::find(base_names_local.begin(), base_names_local.end(), exp_name);
-      if (it != base_names_local.end()) {
-        const auto idx =
-            static_cast<size_t>(std::distance(base_names_local.begin(), it));
-        if (idx < combined_types.size()) {
-          combined_types[idx] = exp_type;
-        }
-      } else {
-        combined_types.push_back(exp_type);
-      }
-    }
-  }
-
-  if (all_columns || active_column_ids.empty()) {
-    return combined_types;
-  }
-
-  std::vector<duckdb::LogicalType> active_result_types;
-  for (auto &column_id : active_column_ids) {
-    if (duckdb::IsRowIdColumnId(column_id)) {
-      continue;
-    }
-    if (column_id < combined_types.size()) {
-      active_result_types.push_back(combined_types[column_id]);
-    }
-  }
-
-  return active_result_types;
-}
-
-// ============================================================================
-// FetchNextResult Helper Methods
-// ============================================================================
-
-void ODataReadBindData::EnsureInitialized() {
-    // Ensure input parameters are set on client before any request
-    if (!input_parameters.empty()) {
-        odata_client->SetInputParameters(input_parameters);
-    }
-
-    // Make sure first page is prefetched once and buffered
-    if (!first_page_cached_) {
-        PrefetchFirstPage();
-    }
-}
-
-SchemaInfo ODataReadBindData::PrepareSchemaInfo() {
-    SchemaInfo info;
-    
-    // Always use full schema for buffering; projection is applied at emission
-    info.all_result_names = GetResultNames(true);
-    info.all_result_types = GetResultTypes(true);
-    info.has_expand = HasExpandedData();
-    
-    // Build name->index map for robust lookups
-    info.name_to_index.reserve(info.all_result_names.size());
-    for (duckdb::idx_t k = 0; k < (duckdb::idx_t)info.all_result_names.size(); ++k) {
-        info.name_to_index.emplace(info.all_result_names[k], k);
-    }
-    
-    return info;
-}
-
-void ODataReadBindData::FetchAdditionalPagesIfNeeded(
-    duckdb::optional_ptr<duckdb::ClientContext> context,
-    const SchemaInfo& schema_info) {
-    const idx_t target = STANDARD_VECTOR_SIZE;
-    idx_t pages_fetched = 0;
-
-    // Fetch additional pages until we have enough buffered rows to fill the
-    // vector or no more pages
-    while (row_buffer->Size() < target && row_buffer->HasNextPage()) {
-        // Ctrl-C and statement timeouts both arrive as an interrupt flag on the
-        // client context; a paging loop that never observes it is unstoppable.
-        if (context != nullptr && context->interrupted) {
-            ERPL_TRACE_INFO("ODATA_READ_BIND", "Interrupted while fetching additional pages");
-            throw duckdb::InterruptException();
-        }
-
-        // Only yield once there is something to emit: DuckDB ends a table scan
-        // on the first empty chunk, so breaking with an empty buffer would
-        // silently truncate the result. A run of empty pages is bounded by the
-        // interrupt check above and by the client's page ceiling instead.
-        if (pages_fetched >= MAX_PAGES_PER_SCAN_CALL && row_buffer->Size() > 0) {
-            ERPL_TRACE_DEBUG(
-                "ODATA_READ_BIND",
-                duckdb::StringUtil::Format(
-                    "Reached the per-scan-call page cap of %llu; emitting a short vector and "
-                    "resuming on the next call",
-                    (unsigned long long)MAX_PAGES_PER_SCAN_CALL));
-            break;
-        }
-
-        auto next_response = odata_client->Get(true);
-        pages_fetched++;
-        if (!next_response) {
-            row_buffer->SetHasNextPage(false);
-            break;
-        }
-        
-        ProcessPageResponse(next_response, schema_info);
-    }
-}
-
-void ODataReadBindData::ProcessPageResponse(
-    std::shared_ptr<ODataEntitySetResponse> response, 
-    const SchemaInfo& schema_info) {
-    
-    // Capture total count once for progress
-    if (response->GetODataVersion() == ODataVersion::V4) {
-        auto total = response->Content()->TotalCount();
-        if (total.has_value() && (!progress_tracker->HasTotalCount() ||
-                                  progress_tracker->GetTotalCount() == 0)) {
-            progress_tracker->SetTotalCount(total.value());
-        }
-    }
-    
-    // Extract expanded data from subsequent pages as well
-    if (schema_info.has_expand && !data_extractor->GetExpandedDataSchema().empty()) {
-        try {
-            ERPL_TRACE_DEBUG("ODATA_READ_BIND",
-                           "Extracting expanded data from subsequent page");
-            data_extractor->ExtractExpandedDataFromResponse(response->RawContent());
-        } catch (const StrictTypingViolation &) {
-            // strict_typing must fail the query, not be downgraded to a warning.
-            throw;
-        } catch (const std::exception &e) {
-            ERPL_TRACE_WARN(
-                "ODATA_READ_BIND",
-                "Failed to extract expanded data from subsequent page: " +
-                    std::string(e.what()));
-        }
-    }
-
-    // Buffer full rows using full schema to keep indices stable
-    auto column_names = schema_info.all_result_names;
-    auto column_types = schema_info.all_result_types;
-    response->Content()->SetConversionFailureLog(conversion_failure_log);
-    auto page_rows = response->ToRows(column_names, column_types);
-    const auto row_count = page_rows.size();
-    row_buffer->AddRows(std::move(page_rows));
-    row_buffer->SetHasNextPage(response->NextUrl().has_value());
-    progress_tracker->IncrementRowsFetched(row_count);
-}
-
-idx_t ODataReadBindData::EmitRowsToOutput(duckdb::DataChunk &output, const SchemaInfo& schema_info) {
-    const idx_t target = STANDARD_VECTOR_SIZE;
-    idx_t to_emit = std::min<idx_t>(row_buffer->Size(), target);
-    
-    for (idx_t i = 0; i < to_emit; i++) {
-        const auto &row = row_buffer->GetNextRow();
-        EmitSingleRowToOutput(output, row, i, schema_info);
-        emitted_row_index_++;
-    }
-    
-    output.SetCardinality(to_emit);
-    return to_emit;
-}
-
-void ODataReadBindData::EmitSingleRowToOutput(
-    duckdb::DataChunk &output, 
-    const std::vector<duckdb::Value> &row, 
-    idx_t row_index,
-    const SchemaInfo& schema_info) {
-    
-    auto null_value = duckdb::Value();
-    
-    idx_t data_slot = 0;
-    for (idx_t j = 0; j < output.ColumnCount(); j++) {
-        if (j < output_column_is_row_id.size() && output_column_is_row_id[j]) {
-            // A synthetic row id: monotonically increasing over the scan, which is all a
-            // row id has to be here. Emitting a data value would be a type mismatch.
-            output.SetValue(j, row_index,
-                            duckdb::Value::BIGINT(static_cast<int64_t>(emitted_row_index_)));
-            continue;
-        }
-
-        duckdb::idx_t original_column_index = GetOriginalColumnIndex(data_slot);
-        data_slot++;
-
-        if (original_column_index < schema_info.all_result_names.size()) {
-            auto value = GetColumnValue(original_column_index, row, schema_info);
-            output.SetValue(j, row_index, value);
-        } else {
-            auto col_type = output.data[j].GetType();
-            output.SetValue(j, row_index, null_value.DefaultCastAs(col_type));
-        }
-    }
-}
-
-duckdb::idx_t ODataReadBindData::GetOriginalColumnIndex(idx_t activated_column_index) const {
-    return (activated_column_index < activated_to_original_mapping.size())
-               ? activated_to_original_mapping[activated_column_index]
-               : activated_column_index;
-}
-
-duckdb::Value ODataReadBindData::GetColumnValue(
-    duckdb::idx_t original_column_index,
-    const std::vector<duckdb::Value> &row,
-    const SchemaInfo& schema_info) {
-    
-    auto null_value = duckdb::Value();
-    
-    // Check if this original index points to an expanded column
-    if (IsExpandedColumn(original_column_index, schema_info)) {
-        return GetExpandedColumnValue(original_column_index, schema_info);
-    } else {
-        return GetRegularColumnValue(original_column_index, row, schema_info);
-    }
-}
-
-bool ODataReadBindData::IsExpandedColumn(duckdb::idx_t original_column_index, const SchemaInfo& schema_info) const {
-    return schema_info.has_expand && 
-           !data_extractor->GetExpandedDataSchema().empty() &&
-           original_column_index >= (schema_info.all_result_names.size() - 
-                                   data_extractor->GetExpandedDataSchema().size());
-}
-
-duckdb::Value ODataReadBindData::GetExpandedColumnValue(
-    duckdb::idx_t original_column_index, 
-    const SchemaInfo& schema_info) {
-    
-    auto null_value = duckdb::Value();
-    
-    size_t expand_index = original_column_index - 
-                         (schema_info.all_result_names.size() - 
-                          data_extractor->GetExpandedDataSchema().size());
-    
-    if (expand_index < data_extractor->GetExpandedDataSchema().size()) {
-        std::string expand_path = data_extractor->GetExpandedDataSchema()[expand_index];
-        
-        // Use global emitted_row_index_ to align with how cache was filled (per input row)
-        auto expand_data = data_extractor->ExtractExpandedDataForRow(
-            std::to_string(emitted_row_index_), expand_path);
-        
-        if (!expand_data.IsNull()) {
-            ERPL_TRACE_DEBUG("ODATA_SCAN", duckdb::StringUtil::Format("Setting expanded data for path '%s'", expand_path.c_str()));
-            return expand_data;
-        } else {
-            ERPL_TRACE_DEBUG("ODATA_SCAN", duckdb::StringUtil::Format("Setting expanded data to NULL for path '%s'", expand_path.c_str()));
-            return null_value;
-        }
-    }
-    
-    return null_value;
-}
-
-duckdb::Value ODataReadBindData::GetRegularColumnValue(
-    duckdb::idx_t original_column_index,
-    const std::vector<duckdb::Value> &row,
-    const SchemaInfo& schema_info) {
-    
-    auto null_value = duckdb::Value();
-    
-    // Regular base column from buffered full row: resolve by name to guard against index drift
-    const std::string &col_name = schema_info.all_result_names[original_column_index];
-    auto it = schema_info.name_to_index.find(col_name);
-    
-    if (it != schema_info.name_to_index.end() && it->second < row.size()) {
-        return row[it->second];
-    } else {
-        return null_value;
-    }
-}
-
-void ODataReadBindData::UpdateProgressTracking(idx_t rows_emitted) {
-    progress_tracker->IncrementRowsFetched(rows_emitted);
-    
-    if (progress_tracker->HasTotalCount() && progress_tracker->GetTotalCount() > 0) {
-        double pct = 100.0 * (double)progress_tracker->GetRowsFetched() /
-                     (double)progress_tracker->GetTotalCount();
-        if (pct > 100.0) {
-            pct = 100.0;
-        }
-        
-        ERPL_TRACE_INFO("ODATA_SCAN",
-                        duckdb::StringUtil::Format(
-                            "Progress: %.2f%% (%llu/%llu)", pct,
-                            (unsigned long long)progress_tracker->GetRowsFetched(),
-                            (unsigned long long)progress_tracker->GetTotalCount()));
-    }
-}
-
-unsigned int ODataReadBindData::FetchNextResult(duckdb::DataChunk &output) {
-    return FetchNextResultInternal(nullptr, output);
-}
-
-unsigned int ODataReadBindData::FetchNextResult(duckdb::ClientContext &context,
-                                                duckdb::DataChunk &output) {
-    return FetchNextResultInternal(&context, output);
-}
-
-unsigned int ODataReadBindData::FetchNextResultInternal(
-    duckdb::optional_ptr<duckdb::ClientContext> context,
-    duckdb::DataChunk &output) {
-    EnsureInitialized();
-
-    auto schema_info = PrepareSchemaInfo();
-    FetchAdditionalPagesIfNeeded(context, schema_info);
-
-    idx_t rows_emitted = EmitRowsToOutput(output, schema_info);
-    UpdateProgressTracking(rows_emitted);
-    
-    return rows_emitted;
-}
-
-bool ODataReadBindData::HasMoreResults() {
-    // If buffer still has rows, we have results to emit
-    if (row_buffer->HasMoreRows()) {
-        return true;
-    }
-    // If first page hasn't been cached yet, we need to deliver it
-    if (!first_page_cached_) {
-        return true;
-    }
-    // Otherwise, only if server indicated a next page
-    return row_buffer->HasNextPage();
-}
-
-void ODataReadBindData::ActivateColumns(
-    const std::vector<duckdb::column_t> &column_ids) {
-    std::stringstream column_ids_str;
-    for (auto &column_id : column_ids) {
-        column_ids_str << column_id << ", ";
-    }
-  ERPL_TRACE_DEBUG("ODATA_READ_BIND",
-                   duckdb::StringUtil::Format("Activating columns: %s",
-                                              column_ids_str.str().c_str()));
-    
-    // Row ids are not data columns, so they take no part in $select - but they DO occupy a
-    // slot in the output chunk, so which slots they are must be remembered. COUNT(*) asks
-    // for nothing but a row id; forgetting that made the emit path fall back to "output
-    // slot j is data column j" and write the first column's string into a BIGINT row-id
-    // vector. See GitHub #132.
-    std::vector<duckdb::column_t> visible_ids;
-    visible_ids.reserve(column_ids.size());
-    output_column_is_row_id.assign(column_ids.size(), false);
-    for (size_t output_index = 0; output_index < column_ids.size(); ++output_index) {
-        const auto column_id = column_ids[output_index];
-        if (duckdb::IsRowIdColumnId(column_id)) {
-            output_column_is_row_id[output_index] = true;
-      ERPL_TRACE_DEBUG("ODATA_READ_BIND",
-                       "Skipping ROW_ID column from activation mapping");
-            continue;
-        }
-        visible_ids.push_back(column_id);
-    }
-
-    active_column_ids = visible_ids;
-    
-    // Build the mapping from activated column index to original column name index
-    activated_to_original_mapping.clear();
-    activated_to_original_mapping.resize(visible_ids.size());
-    
-    for (size_t i = 0; i < visible_ids.size(); ++i) {
-        activated_to_original_mapping[i] = visible_ids[i];
-    ERPL_TRACE_DEBUG("ODATA_READ_BIND",
-                     duckdb::StringUtil::Format(
-                         "Mapping activated index %d to original index %d",
-                         (int)i, (int)visible_ids[i]));
-  }
-
-  // Preserve existing predicate pushdown state (top/skip/expand/filter) across
-  // activation Do NOT reset the helper; just update the column selection
-  ERPL_TRACE_DEBUG("ODATA_READ_BIND",
-                   std::string("ActivateColumns: service_root_mode_ = ") +
-                       (service_root_mode_ ? "true" : "false"));
-  if (!service_root_mode_) {
-    // Refresh the helper's view of the schema before it builds $select: the
-    // expand clause processed after helper creation can add columns, and the
-    // types decide whether $select is safe at all (GitHub #86).
-    PredicatePushdownHelper()->SetColumnSchema(GetResultNames(true),
-                                               GetResultTypes(true));
-    PredicatePushdownHelper()->ConsumeColumnSelection(visible_ids);
-    ERPL_TRACE_DEBUG("ODATA_READ_BIND",
-                     duckdb::StringUtil::Format(
-                         "Select clause: %s",
-                         PredicatePushdownHelper()->SelectClause().c_str()));
-  } else {
-    ERPL_TRACE_DEBUG(
-        "ODATA_READ_BIND",
-        "Service-root mode: skipping predicate pushdown helper creation");
-  }
-}
-
-void ODataReadBindData::AddFilters(
-    const duckdb::optional_ptr<duckdb::TableFilterSet> &filters) {
-  if (service_root_mode_) {
-    ERPL_TRACE_DEBUG(
-        "ODATA_READ_BIND",
-        "Service-root mode: ignoring filters for predicate pushdown");
-    return;
-  }
-    if (filters && !filters->filters.empty()) {
-        std::stringstream filters_str;
-    for (auto &[projected_column_idx, filter] : filters->filters) {
-      filters_str << "Column " << projected_column_idx << ": "
-                  << filter->DebugToString() << std::endl;
-    }
-    ERPL_TRACE_DEBUG("ODATA_READ_BIND",
-                     duckdb::StringUtil::Format("Adding %d filters: %s",
-                                                filters->filters.size(),
-                                                filters_str.str().c_str()));
-        
-        PredicatePushdownHelper()->ConsumeFilters(filters);
-    ERPL_TRACE_DEBUG("ODATA_READ_BIND",
-                     duckdb::StringUtil::Format(
-                         "Filter clause: %s",
-                         PredicatePushdownHelper()->FilterClause().c_str()));
-    } else {
-        ERPL_TRACE_DEBUG("ODATA_READ_BIND", "No filters to add");
-    }
-}
-
-void ODataReadBindData::AddResultModifiers(
-    const std::vector<duckdb::unique_ptr<duckdb::BoundResultModifier>>
-        &modifiers) {
-  // Deliberately does nothing (GitHub #90). DuckDB does not hand table
-  // functions their bound result modifiers, so a SQL `LIMIT`/`OFFSET` never
-  // reaches this extension and is NEVER translated into `$top`/`$skip`:
-  // `SELECT * FROM odata_read(...) LIMIT 10` still pulls every page the
-  // service offers. `$top`/`$skip` are produced only from the explicit
-  // `top=`/`skip=` named parameters (see ODataReadBindHelpers::
-  // ProcessNamedParameters) and their Datasphere/SAC equivalents.
-  //
-  // The code that used to walk the modifiers and synthesise those clauses has
-  // been deleted because it made the read path look as though LIMIT pushdown
-  // existed. This stub survives only so the (equally uncalled)
-  // OdpODataReadBindData::AddResultModifiers keeps linking; delete both
-  // together.
-  (void)modifiers;
-  ERPL_TRACE_DEBUG("ODATA_READ_BIND",
-                   "AddResultModifiers is a no-op: LIMIT is not pushed to $top");
-}
-
-void ODataReadBindData::UpdateUrlFromPredicatePushdown() {
-  if (service_root_mode_) {
-    ERPL_TRACE_DEBUG(
-        "ODATA_READ_BIND",
-        "Service-root mode: skipping URL update from predicate pushdown");
-    return;
-  }
-    ERPL_TRACE_DEBUG("ODATA_READ_BIND", "Updating URL from predicate pushdown");
-    ERPL_TRACE_DEBUG("ODATA_READ_BIND", "Original URL: " + odata_client->Url());
-    
-    auto http_client = odata_client->GetHttpClient();
-    auto auth_params = odata_client->AuthParams();
-    auto prev_url_str = odata_client->Url();
-  auto updated_url =
-      PredicatePushdownHelper()->ApplyFiltersToUrl(odata_client->Url());
-    
-    ERPL_TRACE_DEBUG("ODATA_READ_BIND", "Updated URL: " + updated_url.ToString());
-    
-    // Store the current OData version before creating new client
-    auto current_version = odata_client->GetODataVersion();
-
-  odata_client = std::make_shared<ODataEntitySetClient>(
-      http_client, updated_url, auth_params);
-    
-    // Preserve the OData version to avoid metadata fetching
-    if (current_version != ODataVersion::UNKNOWN) {
-        odata_client->SetODataVersionDirectly(current_version);
-    ERPL_TRACE_DEBUG("ODATA_READ_BIND",
-                     duckdb::StringUtil::Format(
-                         "Preserved OData version %s on new client",
-                         current_version == ODataVersion::V2 ? "V2" : "V4"));
-    }
-
-  // If the finalized URL changed compared to the prefetched one, discard
-  // buffered data so we don't emit unfiltered/unprojected rows. The scan init
-  // will prefetch again.
-    if (first_page_cached_ && prev_url_str != updated_url.ToString()) {
-    ERPL_TRACE_INFO("ODATA_READ_BIND",
-                    "Final URL changed after predicate pushdown; discarding "
-                    "prefetched buffer and caches");
-        if (row_buffer) {
-            row_buffer->Clear();
-            row_buffer->SetHasNextPage(false);
-        }
-        if (progress_tracker) {
-            progress_tracker->Reset();
-        }
-        if (data_extractor) {
-            data_extractor->ClearCache();
-        }
-        emitted_row_index_ = 0;
-        first_page_cached_ = false;
-    }
-}
-
-std::shared_ptr<ODataPredicatePushdownHelper>
-ODataReadBindData::PredicatePushdownHelper() {
-    if (predicate_pushdown_helper == nullptr) {
-    ERPL_TRACE_DEBUG("ODATA_READ_BIND",
-                     "Creating new predicate pushdown helper");
-        
-    // Feed the helper the full resolved schema (metadata order, plus expanded
-    // columns) together with its types. The types let it decide whether
-    // $select is safe rather than guessing from column names (GitHub #86),
-    // and the order matches what DuckDB's column_ids index into.
-    auto column_names = GetResultNames(true);
-    auto column_types = GetResultTypes(true);
-    ERPL_TRACE_DEBUG("ODATA_READ_BIND",
-                     duckdb::StringUtil::Format(
-                         "Creating predicate pushdown helper with %d columns",
-                         (int)column_names.size()));
-    predicate_pushdown_helper =
-        std::make_shared<ODataPredicatePushdownHelper>(column_names, column_types);
-        
-        // Set the OData version for proper filter syntax generation
-        auto odata_version = odata_client->GetODataVersion();
-        predicate_pushdown_helper->SetODataVersion(odata_version);
-    ERPL_TRACE_DEBUG(
-        "ODATA_READ_BIND",
-        "Set OData version " +
-            std::string(odata_version == ODataVersion::V2 ? "V2" : "V4") +
-            " on predicate pushdown helper");
-
-    BindPredicateColumnResolver();
-    }
-
-    return predicate_pushdown_helper;
-}
-
-void ODataReadBindData::BindPredicateColumnResolver() {
-  if (!predicate_pushdown_helper) {
-    return;
-  }
-
-  // DuckDB passes the OUTPUT position (index within the projected column list
-  // from InitGlobal). Translate that through activated_to_original_mapping to
-  // the full-schema index before looking up the column name, because the
-  // activated column order may differ from the schema order (e.g.
-  // column_ids=[8,0] puts Country at output pos 0).
-  predicate_pushdown_helper->SetColumnNameResolver(
-      [this](duckdb::column_t output_index) -> std::string {
-        size_t schema_index = output_index;
-        if (output_index < this->activated_to_original_mapping.size()) {
-          schema_index = this->activated_to_original_mapping[output_index];
-        }
-
-        this->EnsureBaseSchemaResolved();
-        if (schema_index < this->base_result_names.size()) {
-          return this->base_result_names[schema_index];
-        }
-        ERPL_TRACE_ERROR(
-            "ODATA_READ_BIND",
-            duckdb::StringUtil::Format(
-                "Schema column index %d (from output index %d) is out of "
-                "bounds for result names",
-                (int)schema_index, (int)output_index));
-        return std::string();
-      });
-}
-
-duckdb::unique_ptr<ODataReadBindData> ODataReadBindData::CloneForScan() const {
-  // The clone owns every piece of mutable scan state; the source keeps only
-  // the schema and configuration settled during bind (GitHub #75).
-  auto clone = duckdb::make_uniq<ODataReadBindData>(odata_client, true);
-
-  clone->service_root_mode_ = service_root_mode_;
-  clone->InitializeComponents(service_root_mode_);
-
-  // InitializeComponents mints a fresh ConversionFailureLog for this execution, which is
-  // what gives the log the right lifetime - but it also resets strict typing, so the
-  // bind-time choice has to be re-applied or strict_typing=true silently does nothing.
-  clone->SetStrictTyping(strict_typing_);
-
-  // Schema and configuration: pure values, safe to copy.
-  clone->all_result_names = all_result_names;
-  clone->all_result_types = all_result_types;
-  clone->extracted_column_names = extracted_column_names;
-  clone->base_result_names = base_result_names;
-  clone->base_result_types = base_result_types;
-  clone->base_schema_resolved_ = base_schema_resolved_;
-  clone->input_parameters = input_parameters;
-  clone->expand_clause = expand_clause;
-  clone->has_expanded_data = has_expanded_data;
-  clone->active_column_ids = active_column_ids;
-  clone->activated_to_original_mapping = activated_to_original_mapping;
-
-  // The clone gets its own extractor so its per-row expand cache starts empty,
-  // but it must carry the expand schema and the types inferred during bind.
-  if (data_extractor && clone->data_extractor) {
-    clone->data_extractor->SetExpandedDataSchema(
-        data_extractor->GetExpandedDataSchema());
-    clone->data_extractor->SetNestedExpandPaths(
-        data_extractor->GetNestedExpandPaths());
-    const auto &expanded_types = data_extractor->GetExpandedDataTypes();
-    for (size_t i = 0; i < expanded_types.size(); ++i) {
-      clone->data_extractor->UpdateExpandedColumnType(i, expanded_types[i]);
-    }
-  }
-
-  // Carry the clauses generated at bind time (top/skip/expand/count from named
-  // parameters) but re-point the column resolver at the clone.
-  if (predicate_pushdown_helper) {
-    clone->predicate_pushdown_helper =
-        std::make_shared<ODataPredicatePushdownHelper>(*predicate_pushdown_helper);
-    clone->BindPredicateColumnResolver();
-  }
-
-  // Carry rows buffered during bind. The ODP path hands a pre-fetched first
-  // page in via FromEntitySetClient and the service-root path pre-buffers its
-  // synthetic rows; dropping those here would make every execution re-issue a
-  // bare GET (which SAP ODP answers with the entire dataset).
-  if (row_buffer && clone->row_buffer) {
-    clone->row_buffer->AddRows(row_buffer->CopyRows());
-    clone->row_buffer->SetHasNextPage(row_buffer->HasNextPage());
-  }
-  clone->first_page_cached_ = first_page_cached_;
-
-  ERPL_TRACE_DEBUG(
-      "ODATA_READ_BIND",
-      duckdb::StringUtil::Format(
-          "Cloned bind data for scan: %d buffered rows, first_page_cached=%s",
-          (int)(row_buffer ? row_buffer->Size() : 0),
-          first_page_cached_ ? "true" : "false"));
-
-  return clone;
-}
-
-double ODataReadBindData::GetProgressFraction() const {
-    return progress_tracker->GetProgressFraction();
-}
-
-void ODataReadBindData::PrefetchFirstPage() {
-  ERPL_TRACE_DEBUG("ODATA_READ_BIND",
-                   std::string("PrefetchFirstPage: service_root_mode_ = ") +
-                       (service_root_mode_ ? "true" : "false"));
-    if (first_page_cached_) {
-    ERPL_TRACE_DEBUG("ODATA_READ_BIND",
-                     "PrefetchFirstPage: first page already cached, skipping");
-        return;
-    }
-  if (service_root_mode_) {
-    ERPL_TRACE_DEBUG("ODATA_READ_BIND",
-                     "Service-root mode: prefetch skipped (rows prebuffered)");
-    first_page_cached_ = true;
-    row_buffer->SetHasNextPage(false);
-    return;
-  }
-  ERPL_TRACE_DEBUG("ODATA_READ_BIND",
-                   "PrefetchFirstPage: starting first page fetch");
-
-    // Ensure input parameters are set on client before any request
-    if (!input_parameters.empty()) {
-        odata_client->SetInputParameters(input_parameters);
-    }
-
-    auto response = odata_client->Get();
-    BufferFirstPageFromResponse(response);
-
-  ERPL_TRACE_DEBUG("ODATA_READ_BIND",
-                   duckdb::StringUtil::Format(
-        "PrefetchFirstPage: buffered %d rows, has_next_page=%s",
-                       (int)row_buffer->Size(),
-                       row_buffer->HasNextPage() ? "true" : "false"));
-}
-
-void ODataReadBindData::BufferFirstPageFromResponse(
-    std::shared_ptr<ODataEntitySetResponse> response) {
-    if (!response) {
-        ERPL_TRACE_WARN("ODATA_READ_BIND",
-                        "BufferFirstPageFromResponse: no response received");
-        first_page_cached_ = true;
-        row_buffer->SetHasNextPage(false);
-        return;
-    }
-
-    // Extract expanded data if we have expand clauses
-    if (HasExpandedData() && !data_extractor->GetExpandedDataSchema().empty()) {
-        try {
-            ERPL_TRACE_DEBUG("ODATA_READ_BIND",
-                             "Extracting expanded data from buffered first page");
-            data_extractor->ExtractExpandedDataFromResponse(response->RawContent());
-        } catch (const StrictTypingViolation &) {
-            // strict_typing must fail the query, not be downgraded to a warning.
-            throw;
-        } catch (const std::exception &e) {
-            ERPL_TRACE_WARN("ODATA_READ_BIND",
-                            std::string("Failed to extract expanded data from "
-                                        "buffered first page: ") +
-                                e.what());
-        }
-    }
-
-    // Capture total count once for progress (v4 only, when available)
-    if (response->GetODataVersion() == ODataVersion::V4) {
-        auto total = response->Content()->TotalCount();
-        if (total.has_value()) {
-            progress_tracker->SetTotalCount(total.value());
-            ERPL_TRACE_INFO(
-                "ODATA_READ_BIND",
-                duckdb::StringUtil::Format(
-                    "Service reported total row count: %llu",
-                    (unsigned long long)progress_tracker->GetTotalCount()));
-        }
-    }
-
-    // Buffer full schema rows to keep indices stable across projections
-    auto all_result_names_local = GetResultNames(true);
-    auto all_result_types_local = GetResultTypes(true);
-
-    response->Content()->SetConversionFailureLog(conversion_failure_log);
-    auto page_rows = response->ToRows(all_result_names_local, all_result_types_local);
-    row_buffer->AddRows(std::move(page_rows));
-    row_buffer->SetHasNextPage(response->NextUrl().has_value());
-    first_page_cached_ = true;
-}
-
-void ODataReadBindData::SetExtractedColumnNames(
-    const std::vector<std::string> &column_names) {
-    extracted_column_names = column_names;
-    base_schema_resolved_ = false;
-  ERPL_TRACE_DEBUG("ODATA_READ_BIND", "Stored " +
-                                          std::to_string(column_names.size()) +
-                                          " extracted column names");
-}
-
-std::string ODataReadBindData::GetOriginalColumnName(
-    duckdb::column_t activated_column_index) const {
-    if (activated_column_index >= activated_to_original_mapping.size()) {
-    // This is not an error - it's a column that wasn't activated (not selected
-    // by user) Just return empty string for columns we don't have mapping for
-        return "";
-    }
-
-    auto original_index = activated_to_original_mapping[activated_column_index];
-
-  // base_result_names is the reconciled schema: EDMX order first, JSON-only
-  // columns appended. That is exactly the order DuckDB's column indices refer
-  // to, so no other list may be consulted here (GitHub #88).
-  if (original_index >= base_result_names.size()) {
-    ERPL_TRACE_ERROR(
-        "ODATA_READ_BIND",
-        duckdb::StringUtil::Format(
-            "Original column index %d is out of bounds for result names (%d total)",
-            original_index, (int)base_result_names.size()));
-    return "";
-  }
-
-  const auto column_name = base_result_names[original_index];
-  ERPL_TRACE_DEBUG("ODATA_READ_BIND",
-                   duckdb::StringUtil::Format(
-                       "Activated index %d maps to original index %d which "
-                       "is column '%s'",
-                       (int)activated_column_index, (int)original_index,
-                       column_name.c_str()));
-  return column_name;
-}
-
-void ODataReadBindData::SetInputParameters(
-    const std::map<std::string, std::string> &input_params) {
-    input_parameters = input_params;
-  ERPL_TRACE_DEBUG("ODATA_READ_BIND", "Stored " +
-                                          std::to_string(input_params.size()) +
-                                          " input parameters");
-}
-
-const std::map<std::string, std::string> &
-ODataReadBindData::GetInputParameters() const {
-    return input_parameters;
-}
-
-std::shared_ptr<ODataEntitySetClient>
-ODataReadBindData::GetODataClient() const {
-    return odata_client;
-}
-
-void ODataReadBindData::SetExpandClause(const std::string &expand_clause) {
-    this->expand_clause = expand_clause;
-}
-
-std::string ODataReadBindData::GetExpandClause() const { return expand_clause; }
-
-void ODataReadBindData::SetExpandedDataSchema(
-    const std::vector<std::string> &expand_paths) {
-  if (data_extractor) {
-    data_extractor->SetExpandedDataSchema(expand_paths);
-    ERPL_TRACE_INFO("ODATA_READ_BIND", "Set expanded data schema with " +
-                                           std::to_string(expand_paths.size()) +
-                                           " navigation properties");
-  } else {
-    ERPL_TRACE_DEBUG("ODATA_READ_BIND",
-                     "Skipping SetExpandedDataSchema in service root mode");
-  }
-}
-
-// duplicate definition removed; implementation is at top of file to ensure
-// availability before use
-
-bool ODataReadBindData::HasExpandedData() const {
-  return data_extractor && data_extractor->HasExpandedData();
-}
-
-void ODataReadBindData::UpdateExpandedColumnType(
-    const std::string &expand_path, const duckdb::LogicalType &new_type) {
-  if (!HasExpandedData() || !data_extractor) {
-    return;
-  }
-
-  const auto &exp_schema = data_extractor->GetExpandedDataSchema();
-  for (size_t i = 0; i < exp_schema.size(); ++i) {
-    if (exp_schema[i] != expand_path) {
-      continue;
-    }
-    // GetResultTypes() merges the extractor's expanded types over the base
-    // schema by column name on every call, so the extractor is the single
-    // place this type needs to live.
-    data_extractor->UpdateExpandedColumnType(i, new_type);
-    return;
-  }
-}
 
 // -------------------------------------------------------------------------------------------------
 
@@ -2074,49 +205,6 @@ void SetupSchemaFromProbeResult(
 
 } // namespace ODataReadBindHelpers
 
-// ============================================================================
-// Shared Error Handling Utilities
-// ============================================================================
-
-namespace ODataErrorHandling {
-
-/**
- * @brief Convert runtime errors (especially HTTP errors) to user-friendly InvalidInputException
- * 
- * @param e The runtime error to convert
- * @param url The URL that caused the error
- * @param service_type Type of service ("OData" or "ODP OData") for error messages
- * @param discovery_function Name of discovery function to suggest (e.g., "sap_odata_show()" or "sap_odp_odata_show()")
- * @return InvalidInputException with user-friendly error message
- */
-duckdb::InvalidInputException ConvertHttpErrorToUserFriendly(const std::runtime_error& e, 
-                                                           const std::string& url,
-                                                           const std::string& service_type,
-                                                           const std::string& discovery_function) {
-    std::string error_msg = e.what();
-    
-    if (error_msg.find("HTTP 404") != std::string::npos) {
-        return duckdb::InvalidInputException(service_type + " service not found at URL: " + url + 
-            ". Please check if the URL path is correct, especially the entity set name. " +
-            "Use " + discovery_function + " to discover available entity sets.");
-    } else if (error_msg.find("HTTP 401") != std::string::npos) {
-        return duckdb::InvalidInputException("Authentication failed for " + service_type + " service at: " + url + 
-            ". Please check your credentials in the secret.");
-    } else if (error_msg.find("HTTP 403") != std::string::npos) {
-        return duckdb::InvalidInputException("Access forbidden to " + service_type + " service at: " + url + 
-            ". Please check if your user has permission to access this service.");
-    } else if (error_msg.find("Connection failed") != std::string::npos) {
-        return duckdb::InvalidInputException("Failed to connect to " + service_type + " service at: " + url + 
-            ". Please check if the server is running and accessible.");
-    } else {
-        // Re-throw with additional context
-        return duckdb::InvalidInputException("Failed to access " + service_type + " service at: " + url + 
-            ". Error: " + error_msg);
-    }
-}
-
-} // namespace ODataErrorHandling
-
 duckdb::unique_ptr<FunctionData>
 ODataReadBind(ClientContext &context, TableFunctionBindInput &input,
               vector<LogicalType> &return_types, vector<string> &names) {
@@ -2200,88 +288,6 @@ ODataReadBind(ClientContext &context, TableFunctionBindInput &input,
   }
 }
 
-ODataReadGlobalState::ODataReadGlobalState(
-    duckdb::unique_ptr<ODataReadBindData> scan_state)
-    : scan_state(std::move(scan_state)) {
-}
-
-namespace {
-
-// Returns the object that owns the scan state for this execution.
-//
-// odata_read(), the ATTACHed odata_table_scan and the SAC readers use
-// ODataReadGlobalState, so each execution of a bound plan gets its own state.
-// The Datasphere readers reuse ODataReadScan with their own init-global
-// functions, which still return a bare GlobalTableFunctionState; those keep the
-// historical behaviour of scanning straight out of the bind data.
-ODataReadBindData &ResolveScanState(const duckdb::FunctionData *bind_data,
-                                    const GlobalTableFunctionState *global_state) {
-  auto *odata_global = dynamic_cast<const ODataReadGlobalState *>(global_state);
-  if (odata_global != nullptr) {
-    return const_cast<ODataReadGlobalState *>(odata_global)->Scan();
-  }
-  return bind_data->CastNoConst<ODataReadBindData>();
-}
-
-} // namespace
-
-unique_ptr<GlobalTableFunctionState>
-ODataReadTableInitGlobalState(ClientContext &context,
-                              TableFunctionInitInput &input) {
-  auto &bind_data = input.bind_data->CastNoConst<ODataReadBindData>();
-
-  // Projection, filter pushdown and the first-page prefetch all mutate scan
-  // state, so they run against a private clone. The bind data stays exactly as
-  // bind left it, which is what makes a second EXECUTE of the same prepared
-  // statement (and a self-join of one odata_read call) work (GitHub #75).
-  auto scan_state = bind_data.CloneForScan();
-
-  scan_state->ActivateColumns(input.column_ids);
-  scan_state->AddFilters(input.filters);
-  scan_state->UpdateUrlFromPredicatePushdown();
-  // Prefetch first page after URL is finalized so progress can show early and
-  // tiny scans return immediately
-  scan_state->PrefetchFirstPage();
-
-  return duckdb::make_uniq<ODataReadGlobalState>(std::move(scan_state));
-}
-
-double ODataReadTableProgress(ClientContext &, const FunctionData *func_data,
-                              const GlobalTableFunctionState *global_state) {
-  return ResolveScanState(func_data, global_state).GetProgressFraction();
-}
-
-void ODataReadScan(ClientContext &context, TableFunctionInput &data,
-                   DataChunk &output) {
-  // Scan state lives on the per-execution clone owned by the global state, so a
-  // bound plan can be executed more than once (GitHub #75). ResolveScanState
-  // falls back to the bind data for callers that still supply a bare
-  // GlobalTableFunctionState.
-  auto &scan_state =
-      ResolveScanState(data.bind_data.get(), data.global_state.get());
-
-  ERPL_TRACE_DEBUG("ODATA_SCAN", "Starting OData scan operation");
-
-  if (!scan_state.HasMoreResults()) {
-    ERPL_TRACE_DEBUG("ODATA_SCAN", "No more results available");
-    // End of scan: surface the per-column conversion failures once, against the
-    // instance that actually accumulated them (GitHub #74).
-    scan_state.ReportConversionFailures();
-    return;
-  }
-
-  ERPL_TRACE_DEBUG("ODATA_SCAN", "Fetching next result set");
-  // The context reaches the paging loop so a long scan stays interruptible and
-  // yields between pages (GitHub #78).
-  auto rows_fetched = scan_state.FetchNextResult(context, output);
-  ERPL_TRACE_INFO("ODATA_SCAN",
-                  duckdb::StringUtil::Format("Fetched %d rows", rows_fetched));
-
-  if (!scan_state.HasMoreResults()) {
-    scan_state.ReportConversionFailures();
-  }
-}
-
 TableFunctionSet CreateODataReadFunction() {
     TableFunctionSet function_set("odata_read");
     
@@ -2303,6 +309,5 @@ TableFunctionSet CreateODataReadFunction() {
     function_set.AddFunction(read_entity_set);
     return function_set;
 }
-
 
 } // namespace erpl_web
